pyproject.toml
[project]
name = "stocksense"
version = "1.0.1"
description = "Package for value investing and advanced stock analytics"
authors = [
    {name="Francisco Silva", email="francisco3597@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "PyYAML",
    "numpy",
    "polars",
    "loguru",
    "tqdm",
    "streamlit",
    "yfinance",
    "beautifulsoup4",
    "requests_cache",
    "requests_ratelimiter",
    "pyrate_limiter",
    "polars-talib",
    "pygad",
    "xgboost",
    "scikit-learn",
    "plotly",
    "nbformat>=4.2.0",
    "watchdog",
    "ipykernel",
    "shap",
    "pre-commit"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
omit = [
    "*/conftest.py", "*/test*", # omit test files
]

[tool.coverage.report]
fail_under = 80

[tool.setuptools.packages.find]
where = ["stocksense"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[project.scripts]
stocksense-app = "app.home:main"
stocksense = "main:main"
